+++ docs/interactions.rst
@@ -1,0 +1,19 @@
+Interactions
+##########################################
+
+Interactions such as slash commands and buttons are some sexy lil additions to Discord. As such, you'll probably want to use them. Fortunately, VoxelBotUtils makes this pretty easy.
+
+Slash Commands
+------------------------------------------
+
+Slash commands
+
+Components
+------------------------------------------
+
+Components
+
+Buttons
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+
+Buttons
