--- docs/voxelbotutils.cogs.interactions.rst
@@ -1,2 +1,0 @@
-.. _interactionsmodule:
-
