+++ voxelbotutils/cogs/utils/interactions/components/buttons.py
@@ -7,7 +7,6 @@ import re
 import discord
 
 from .models import DisableableComponent
-from ..interaction_messageable import InteractionMessageable
 
 
 class ButtonStyle(enum.IntEnum):
@@ -122,47 +121,3 @@ class Button(DisableableComponent):
             custom_id=data.get("custom_id"),
             emoji=emoji,
         )
-
-
-class ButtonInteractionPayload(InteractionMessageable):
-    """
-    An interaction messageable that comes from a button interaction.
-    """
-
-    __slots__ = ("button", "user", "message", "guild", "channel", "_state", "data")
-    ACK_RESPONSE_TYPE = 6
-    ACK_IS_EDITABLE = False
-    IS_COMPONENT = True
-
-    @classmethod
-    def from_payload(cls, data, state):
-        """
-        Construct a response from the gateway payload.
-        """
-
-        # Reconstruct the button that was clicked
-        clicked_button_id = data['data']['custom_id']
-        clicked_button_payload = None
-        for action_row in data['message']['components']:
-            for button in action_row['components']:
-                if button['custom_id'] == clicked_button_id:
-                    clicked_button_payload = button
-                    break
-            if clicked_button_payload is not None:
-                break
-        clicked_button_object = Button.from_dict(clicked_button_payload)
-
-        # Make the response
-        v = cls()
-        v.data = data
-        v._state = state
-        v.button = clicked_button_object
-        channel, guild = state._get_guild_channel(data)
-        v.channel = channel
-        v.guild = guild
-        v.message = discord.Message(channel=channel, data=data['message'], state=state)
-        if guild:
-            v.user = discord.Member(data=data['member'], guild=guild, state=state)
-        else:
-            v.user = discord.User(data=data['user'], state=state)
-        return v
